Modify sles12 to just use zipl to prepare device for IPL
Previously we use grub2-zipl-setup as part of deploy fba image,
the purpose of this command is it will create initial ramdisk
images and then prepare device for initial problem load via zipl,
actually during deployment we don't need to recreate the initrd
image, but just execute zipl would be enough. Sometime the new
generated initrd is different from the original one(for example,
zypper update sles12sp1), it will cause the updated system fail
to start up , so submit this patch to solve it.

Change-Id: Ib4005e93efcebb82393cde31257ed6743835442f