Update patch set 2
Patch Set 2:

OK, I can append the source file. Just because, I use a audit2allow to generate the binary file directly, no source file are written, so I don't past the source file. I can generate the source .te file and paste it here.

For the question Steve has mentioned, I want to make clarification, this file is just want to built in the a single file that installs the IUCV daemon, which is used by customer manually.

Patch-set: 2