Update patch set 2
Patch Set 2:

What is the input to audit2allow ?  When you retire 50 years from now, someone will need to know what input you provided to the audit2allow command.  :-)  They will not look for a bugzilla update.  I would expect them to look in the code somewhere.  Is there somewhere in the code that you document this ?

Patch-set: 2