Update patch set 5
Patch Set 5: Code-Review+1

(2 comments)

If I understand this correctly then my comments are minor because a customer would not see the readme file but instead they are used to inform future developers.

My understanding is that only the .pp file will be shipped.  The other two parts remain in the code tree for developers to use.  Correct?

Patch-set: 5
Label: Code-Review=+1