Update patch set 1
Patch Set 1:

In my mind this is more for completeness of the error message than because someone is actually going to mess with the file privs.  It just seems good form to put out which file we can't stat.  But if the problem is that the file needs root privs and the user doesn't have that, the user ought to use sudo.

Patch-set: 1